william your friend is psycho

+++ src/Main.java
@@ -19,7 +19,7 @@ public class Main
             Decoder.setMyMapsFromJson(r, s);
 
            Thread.sleep(3000);
-           int directions[] = r.calculateSpeeds();
+           double directions[] = r.calculateSpeeds();
            String speedResult = connection.sendSpeed(directions[0], directions[1]);
            if(directions[0] == 0 && directions[1] == 0)
            {
